% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpipwR-package.R
\docType{package}
\name{EpipwR}
\alias{EpipwR-package}
\alias{EpipwR}
\title{EpipwR: Efficient Power Analysis for EWAS with Continuous or Binary Outcomes}
\description{
EpipwR: Efficient Power Analysis for EWAS with Continuous or Binary Outcomes
}
\details{
A quasi-simulation based approach to performing power analysis
for EWAS (Epigenome-wide association studies) with continuous or binary
outcomes. 'EpipwR' relies on empirical EWAS datasets to determine power
at specific sample sizes while keeping computational cost low. EpipwR can
be run with a variety of standard statistical tests, controlling for either
a false discovery rate or a family-wise type I error rate.

For a detailed overview, please use \code{vignette("EpipwR")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jbarth216/EpipwR}
  \item Report bugs at \url{https://github.com/jbarth216/EpipwR}
}

}
\author{
\strong{Maintainer}: Jackson Barth \email{Jackson_Barth@Baylor.edu} (\href{https://orcid.org/0009-0009-6307-9928}{ORCID})

Authors:
\itemize{
  \item Austin Reynolds
}

Other contributors:
\itemize{
  \item Mary Lauren Benton [contributor]
  \item Carissa Fong [contributor]
}

}
\keyword{internal}
