% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpipwR_plot.R
\name{EpipwR_plot}
\alias{EpipwR_plot}
\title{Plotting EpipwR Output}
\usage{
EpipwR_plot(df)
}
\arguments{
\item{df}{A data frame containing output from the \code{get_power_cont()}
or \code{get_power_cc()} function.}
}
\value{
A line plot with sample size on the x-axis and power on the y-axis.
Multiple lines are plotted for each correlation/effect size mean specified.
}
\description{
Plots 95\% error bars of EpipwR output.
Different sample sizes are displayed along the x-axis with lines
representing distinct correlation means. This functions should
only be used with EpipwR output.
}
\examples{
out <- get_power_cont(100,10000,c(70,80,90,100),.05,c(.3,.4,.5))
EpipwR_plot(out)
}
